Put FastAPI's endpoint functions in a separate module
